Assignment2 -Updated UI to fit new commands. -More exception handling.
TennisDatabase
-Writes all players and matches from the database to a text file.
-Now handles new functionality.
-Updated interface

Completed the TennisPlayersContainer class which implements a Binary Search Tree. New functionality compared to the LinkedList implementation.
-Removes a player.
-Clears all players.
-Condenses all players preorder to a string.
-Creates a specialized iterator of either inorder or preorder.
-Updated interface.

Added functionality to the TennisMatchesContainer class
-Removes all matches of a given player when they are removed from the database.
-Clears all matches.
-Condenses all matches in sorted order to a string.
-Updated interface.

Added an array based Queue which implements a circular array.
-Uses a circular array to avoid right shifting.
-Implements standard queue methods.
-Has an interface.

Added an iterator for the TennisPlayersContainer which uses the QueueArrayBased as its underlying container to iterate through the container in a specified order
-Implements the java iterator interface.
-Supports any given order (TennisPlayersContainer inserts items in either inorder or preorder when the iterator function is called)

Made all classes in the TennisDatabase package packageprivate except for those specified in the assignment